requirements.txt
pytorch-models==0.1.7
tensorboardX==1.4
torch==0.4.1